fork

+++ yudao-module-mall/yudao-module-product-biz/src/main/java/cn/iocoder/yudao/module/product/service/property/ProductPropertyServiceImpl.java
@@ -1,32 +1,28 @@
 package cn.iocoder.yudao.module.product.service.property;
 
+import cn.hutool.core.util.ObjUtil;
 import cn.iocoder.yudao.framework.common.pojo.PageResult;
-import cn.iocoder.yudao.module.product.controller.admin.property.vo.property.ProductPropertyAndValueRespVO;
-import cn.iocoder.yudao.module.product.controller.admin.property.vo.property.*;
-import cn.iocoder.yudao.module.product.controller.admin.property.vo.value.ProductPropertyValueCreateReqVO;
-import cn.iocoder.yudao.module.product.controller.admin.property.vo.value.ProductPropertyValueRespVO;
+import cn.iocoder.yudao.module.product.controller.admin.property.vo.property.ProductPropertyCreateReqVO;
+import cn.iocoder.yudao.module.product.controller.admin.property.vo.property.ProductPropertyListReqVO;
+import cn.iocoder.yudao.module.product.controller.admin.property.vo.property.ProductPropertyPageReqVO;
+import cn.iocoder.yudao.module.product.controller.admin.property.vo.property.ProductPropertyUpdateReqVO;
 import cn.iocoder.yudao.module.product.convert.property.ProductPropertyConvert;
-import cn.iocoder.yudao.module.product.convert.propertyvalue.ProductPropertyValueConvert;
 import cn.iocoder.yudao.module.product.dal.dataobject.property.ProductPropertyDO;
-import cn.iocoder.yudao.module.product.dal.dataobject.property.ProductPropertyValueDO;
 import cn.iocoder.yudao.module.product.dal.mysql.property.ProductPropertyMapper;
-import cn.iocoder.yudao.module.product.dal.mysql.property.ProductPropertyValueMapper;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 
 import javax.annotation.Resource;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 import static cn.iocoder.yudao.framework.common.exception.util.ServiceExceptionUtil.exception;
-import static cn.iocoder.yudao.module.product.enums.ErrorCodeConstants.PROPERTY_NOT_EXISTS;
+import static cn.iocoder.yudao.module.product.enums.ErrorCodeConstants.*;
 
 /**
- * 规格名称 Service 实现类
+ * 商品属性项 Service 实现类
  *
  * @author 芋道源码
  */
@@ -38,20 +34,21 @@ public class ProductPropertyServiceImpl implements ProductPropertyService {
     private ProductPropertyMapper productPropertyMapper;
 
     @Resource
-    private ProductPropertyValueMapper productPropertyValueMapper;
+    @Lazy // 延迟加载，解决循环依赖问题
+    private ProductPropertyValueService productPropertyValueService;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Long createProperty(ProductPropertyCreateReqVO createReqVO) {
+        // 如果已经添加过该属性项，直接返回
+        ProductPropertyDO dbProperty = productPropertyMapper.selectByName(createReqVO.getName());
+        if (dbProperty != null) {
+            return dbProperty.getId();
+        }
+
         // 插入
         ProductPropertyDO property = ProductPropertyConvert.INSTANCE.convert(createReqVO);
         productPropertyMapper.insert(property);
-
-        //插入属性值
-        List<ProductPropertyValueCreateReqVO> propertyValueList = createReqVO.getPropertyValueList();
-        List<ProductPropertyValueDO> productPropertyValueDOList = ProductPropertyValueConvert.INSTANCE.convertList03(propertyValueList);
-        productPropertyValueDOList.forEach(x-> x.setPropertyId(property.getId()));
-        productPropertyValueMapper.insertBatch(productPropertyValueDOList);
         // 返回
         return property.getId();
     }
@@ -59,33 +56,48 @@ public class ProductPropertyServiceImpl implements ProductPropertyService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void updateProperty(ProductPropertyUpdateReqVO updateReqVO) {
-        // 校验存在
-        this.validatePropertyExists(updateReqVO.getId());
+        validatePropertyExists(updateReqVO.getId());
+        // 校验名字重复
+        ProductPropertyDO productPropertyDO = productPropertyMapper.selectByName(updateReqVO.getName());
+        if (productPropertyDO != null &&
+                ObjUtil.notEqual(productPropertyDO.getId(), updateReqVO.getId())) {
+            throw exception(PROPERTY_EXISTS);
+        }
+
         // 更新
         ProductPropertyDO updateObj = ProductPropertyConvert.INSTANCE.convert(updateReqVO);
         productPropertyMapper.updateById(updateObj);
-        //更新属性值，先删后加
-        productPropertyValueMapper.deletePropertyValueByPropertyId(updateReqVO.getId());
-        List<ProductPropertyValueCreateReqVO> propertyValueList = updateReqVO.getPropertyValueList();
-        List<ProductPropertyValueDO> productPropertyValueDOList = ProductPropertyValueConvert.INSTANCE.convertList03(propertyValueList);
-        productPropertyValueDOList.forEach(x-> x.setPropertyId(updateReqVO.getId()));
-        productPropertyValueMapper.insertBatch(productPropertyValueDOList);
     }
 
     @Override
     public void deleteProperty(Long id) {
         // 校验存在
-        this.validatePropertyExists(id);
+        validatePropertyExists(id);
+        // 校验其下是否有规格值
+        if (productPropertyValueService.getPropertyValueCountByPropertyId(id) > 0) {
+            throw exception(PROPERTY_DELETE_FAIL_VALUE_EXISTS);
+        }
+
         // 删除
         productPropertyMapper.deleteById(id);
-        //同步删除属性值
-        productPropertyValueMapper.deletePropertyValueByPropertyId(id);
+        // 同步删除属性值
+        productPropertyValueService.deletePropertyValueByPropertyId(id);
     }
 
     private void validatePropertyExists(Long id) {
         if (productPropertyMapper.selectById(id) == null) {
             throw exception(PROPERTY_NOT_EXISTS);
         }
+    }
+
+    @Override
+    public List<ProductPropertyDO> getPropertyList(ProductPropertyListReqVO listReqVO) {
+        return productPropertyMapper.selectList(listReqVO);
+    }
+
+    @Override
+    public PageResult<ProductPropertyDO> getPropertyPage(ProductPropertyPageReqVO pageReqVO) {
+        return productPropertyMapper.selectPage(pageReqVO);
     }
 
     @Override
@@ -98,54 +110,4 @@ public class ProductPropertyServiceImpl implements ProductPropertyService {
         return productPropertyMapper.selectBatchIds(ids);
     }
 
-    @Override
-    public PageResult<ProductPropertyDO> getPropertyPage(ProductPropertyPageReqVO pageReqVO) {
-        return productPropertyMapper.selectPage(pageReqVO);
-    }
-
-    @Override
-    public PageResult<ProductPropertyAndValueRespVO> getPropertyListPage(ProductPropertyPageReqVO pageReqVO) {
-        //获取属性列表
-        PageResult<ProductPropertyDO> pageResult = productPropertyMapper.selectPage(pageReqVO);
-        PageResult<ProductPropertyAndValueRespVO> propertyRespVOPageResult = ProductPropertyConvert.INSTANCE.convertPage(pageResult);
-        List<Long> propertyIds = propertyRespVOPageResult.getList().stream().map(ProductPropertyAndValueRespVO::getId).collect(Collectors.toList());
-
-        //获取属性值列表
-        List<ProductPropertyValueDO> productPropertyValueDOList = productPropertyValueMapper.getPropertyValueListByPropertyId(propertyIds);
-        List<ProductPropertyValueRespVO> propertyValueRespVOList = ProductPropertyValueConvert.INSTANCE.convertList(productPropertyValueDOList);
-        //组装一对多
-        propertyRespVOPageResult.getList().forEach(x->{
-            Long propertyId = x.getId();
-            List<ProductPropertyValueRespVO> valueDOList = propertyValueRespVOList.stream().filter(v -> v.getPropertyId().equals(propertyId)).collect(Collectors.toList());
-            x.setValues(valueDOList);
-        });
-        return propertyRespVOPageResult;
-    }
-
-    private List<ProductPropertyValueDO> getPropertyValueListByPropertyId(List<Long> propertyIds) {
-        return productPropertyValueMapper.getPropertyValueListByPropertyId(propertyIds);
-    }
-
-    @Override
-    public ProductPropertyAndValueRespVO getPropertyResp(Long id) {
-        //查询规格
-        ProductPropertyDO property = getProperty(id);
-        ProductPropertyAndValueRespVO propertyRespVO = ProductPropertyConvert.INSTANCE.convert(property);
-        //查询属性值
-        List<ProductPropertyValueDO> valueDOList = productPropertyValueMapper.getPropertyValueListByPropertyId(Arrays.asList(id));
-        List<ProductPropertyValueRespVO> propertyValueRespVOS = ProductPropertyValueConvert.INSTANCE.convertList(valueDOList);
-        //组装
-        propertyRespVO.setValues(propertyValueRespVOS);
-        return propertyRespVO;
-    }
-
-    @Override
-    public List<ProductPropertyAndValueRespVO> getPropertyAndValueList(Collection<Long> ids) {
-        List<ProductPropertyAndValueRespVO> productPropertyRespVO = ProductPropertyConvert.INSTANCE.convertList(productPropertyMapper.selectBatchIds(ids));
-        //查询属性值
-        List<ProductPropertyValueDO> valueDOList = productPropertyValueMapper.selectBatchIds(ids);
-        Map<Long, List<ProductPropertyValueDO>> propertyValuesMap = valueDOList.stream().collect(Collectors.groupingBy(ProductPropertyValueDO::getPropertyId));
-        productPropertyRespVO.forEach(p -> p.setValues(ProductPropertyValueConvert.INSTANCE.convertList(propertyValuesMap.get(p.getId()))));
-        return productPropertyRespVO;
-    }
 }
